backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Wayne Hsieh <lakerwayne@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
motor = "^3.5.1"
sse-starlette = "^2.1.3"
langchain-community = "^0.2.16"
langgraph = "^0.2.19"
openai = "^1.44.1"
python-jose = "^3.3.0"
rich = "^13.8.1"
langchain-openai = "0.1.23"
langchain-google-vertexai = "1.0.10"
httpx = "^0.27.2"
qdrant-client = "1.12.1"
requests = "^2.32.3"
bs4 = "^0.0.2"
python-multipart = "^0.0.20"
pypdf = "^5.4.0"
slack-sdk = "^3.27.0"
slack-bolt = "^1.18.1"
aiohttp = "^3.9.3"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
